# pkg/models/forms/inputForms/authInputForms.go
package inputForms

type SignInInput struct {
	Login    string `json:"login" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type SignUpUserForm struct {
	Phone      string `json:"phone" binding:"required"`
	Password   string `json:"password" binding:"required"`
	FirstName  string `json:"first_name" binding:"required"`
	SecondName string `json:"second_name" binding:"required"`
}

type TokenReqBody struct {
	RefreshToken string `json:"refresh_token" binding:"required"`
}